requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
huggingface-hub==0.19.4
python-dotenv==1.0.0
pandas>=1.5.0,<2.1
plotly==5.18.0
dash==2.14.2
dash-bootstrap-components==1.5.0
pytest==7.4.3
httpx==0.25.2
python-multipart==0.0.6
aiofiles==23.2.1
